moved common/models/settings to types/settings
