+++ apps/studio/electron/main/storage/index.ts
@@ -6,7 +6,7 @@ import type {
     ProjectsCache,
     UserMetadata,
     UserSettings,
-} from '/common/models/settings';
+} from '@onlook/types/settings';
 
 export enum StorageType {
     USER_SETTINGS = 'user-settings',
