+++ apps/studio/src/components/Announcement/index.tsx
@@ -7,7 +7,7 @@ import { Input } from '@onlook/ui/input';
 import { Toggle } from '@onlook/ui/toggle';
 import { toast } from '@onlook/ui/use-toast';
 import { Links, MainChannels } from '@onlook/types/constants';
-import type { UserSettings } from '/common/models/settings';
+import type { UserSettings } from '@onlook/types/settings';
 import supabase from '@onlook/supabase/clients';
 import { Icons } from '@onlook/ui/icons';
