+++ apps/studio/src/routes/editor/TopBar/OpenCode/index.tsx
@@ -14,7 +14,7 @@ import { MainChannels } from '@onlook/types/constants';
 import { IDE, IdeType } from '/common/ide';
 import type { WebViewElement } from '/common/models/element';
 import type { TemplateNode } from '/common/models/element/templateNode';
-import type { UserSettings } from '/common/models/settings';
+import type { UserSettings } from '@onlook/types/settings';
 
 const OpenCode = observer(() => {
     const editorEngine = useEditorEngine();
