+++ packages/types/src/settings/index.ts
